Issue #99 UI層の修正： UI層の修正、ビジネスロジック層との分離を強化しました。
