+++ iOSEngineerCodeCheck/View/Screen/ContentView.swift
@@ -1,0 +1,49 @@
+//
+//  ContentView.swift
+//  iOSEngineerCodeCheck
+//
+//  Created by taro-taryo on 2024/11/10.
+// Copyright © 2024 YUMEMI Inc. All rights reserved.
+// Copyright © 2024 taro-taryo. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    @StateObject private var viewModel = RepositoryListViewModel()
+
+    var body: some View {
+        NavigationView {
+            VStack {
+                SearchBar(text: $viewModel.searchText) {
+                    viewModel.searchRepositories()
+                }
+                List(viewModel.repositories) { repository in
+                    NavigationLink(destination: DetailView(repository: repository)) {
+                        RepositoryRow(repository: repository)
+                    }
+                }
+                .navigationTitle("Repositories")
+                .alert(item: $viewModel.error) { error in
+                    Alert(
+                        title: Text("Error"),
+                        message: Text(error.localizedDescription),
+                        dismissButton: .default(Text("OK"))
+                    )
+                }
+                .modifier(KeyboardAvoider())
+            }
+        }
+    }
+}
